seperate state of app
